test: break apart monolithic test file
